+++ backend/rush/src/main/java/rush/rush/controller/group/ManageGroupController.java
@@ -1,4 +1,4 @@
-package rush.rush.controller;
+package rush.rush.controller.group;
 
 import java.net.URI;
 import lombok.RequiredArgsConstructor;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import rush.rush.security.CurrentUser;
 import rush.rush.security.user.UserPrincipal;
-import rush.rush.service.GroupService;
+import rush.rush.service.group.GroupService;
 
 @RequiredArgsConstructor
 @RestController
