+++ backend/rush/src/main/java/rush/rush/service/article/DeleteArticleService.java
@@ -1,4 +1,4 @@
-package rush.rush.service;
+package rush.rush.service.article;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
