+++ backend/rush/src/main/java/rush/rush/service/article/FindArticleService.java
@@ -1,4 +1,4 @@
-package rush.rush.service;
+package rush.rush.service.article;
 
 import java.util.Collections;
 import java.util.List;
