+++ backend/rush/src/main/java/rush/rush/service/article/comment/DeleteCommentService.java
@@ -1,4 +1,4 @@
-package rush.rush.service;
+package rush.rush.service.article.comment;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
