+++ backend/rush/src/main/java/rush/rush/service/group/GroupService.java
@@ -1,4 +1,4 @@
-package rush.rush.service;
+package rush.rush.service.group;
 
 import java.util.List;
 import java.util.Optional;
