+++ backend/rush/src/test/java/rush/rush/service/CreateArticleServiceTest.java
@@ -10,6 +10,7 @@ import rush.rush.domain.AuthProvider;
 import rush.rush.domain.User;
 import rush.rush.dto.CreateArticleRequest;
 import rush.rush.repository.UserRepository;
+import rush.rush.service.article.CreateArticleService;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Transactional
