+++ backend/rush/src/test/java/rush/rush/service/DeleteCommentServiceTest.java
@@ -16,6 +16,7 @@ import rush.rush.domain.User;
 import rush.rush.repository.ArticleRepository;
 import rush.rush.repository.CommentRepository;
 import rush.rush.repository.UserRepository;
+import rush.rush.service.article.comment.DeleteCommentService;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Transactional
